requirements.txt
ipykernel
google-genai
json
cryptography
requests
# requests[socks]
pysocks
numpy
stem
presidio-anonymizer
presidio-analyzer